refactor: rename `fake_target_converters.py` to `plugin_converters.py`
