+++ src/ditto/conversion/__init__.py
@@ -1,2 +1,2 @@
 from .aten_ops_converters import *
-from .fake_target_converters import *
+from .plugin_converters import *
